strip out error handling API into a separate header
